% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courseListing.R
\name{courseListingServer}
\alias{courseListingServer}
\title{courseListingServer}
\usage{
courseListingServer(
  id,
  inSemesterCodes,
  theMasterCourses,
  theFaculty,
  theCoursesOffered,
  theCombinedData,
  theCourseInventory,
  showDebugCues,
  synchronize,
  chosenSemester = NULL
)
}
\arguments{
\item{id}{The id}

\item{inSemesterCodes}{the inSemesterCodes}

\item{theMasterCourses}{the master courses}

\item{theFaculty}{the Faculty}

\item{theCoursesOffered}{the coursesOffered}

\item{theCombinedData}{theCombined Data}

\item{theCourseInventory}{the courseInventory}

\item{showDebugCues}{logical variable to show debug cues}

\item{synchronize}{synchronize across boxes}

\item{chosenSemester}{the semester chosen}
}
\description{
courseListingServer
}
