% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courseListing.R
\name{courseListingUI}
\alias{courseListingUI}
\title{courseListingUI}
\usage{
courseListingUI(id, label = "Course Listing")
}
\arguments{
\item{id}{the id}

\item{label}{the label}
}
\description{
courseListingUI
}
