% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullSchedulingBox.R
\name{fullSchedulingBoxModuleUI}
\alias{fullSchedulingBoxModuleUI}
\title{fullSchedulingBoxModuleUI}
\usage{
fullSchedulingBoxModuleUI(id)
}
\arguments{
\item{id}{The element id}
}
\description{
fullSchedulingBoxModuleUI
}
