% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instructorAssignment.R
\name{instructorAssignmentModuleServer}
\alias{instructorAssignmentModuleServer}
\title{instructorAssignmentModuleServer}
\usage{
instructorAssignmentModuleServer(
  id,
  input,
  output,
  session,
  inSemester,
  theMasterCourses,
  theCombinedData,
  theLeaveData,
  inSemesterCodes,
  synchronize,
  chosenSemester = NULL,
  suspendWhenHidden,
  allowUpdate = FALSE,
  rds.path = NULL
)
}
\arguments{
\item{id}{The id}

\item{input}{The input}

\item{output}{The output}

\item{session}{The session}

\item{inSemester}{The input semester}

\item{theMasterCourses}{The master course list}

\item{theCombinedData}{The combined data}

\item{theLeaveData}{The leave data}

\item{inSemesterCodes}{the semester codes}

\item{synchronize}{synchronize boxes}

\item{chosenSemester}{the chosen semester}

\item{suspendWhenHidden}{suspendWhenHidden}

\item{allowUpdate}{Allow for updates}

\item{rds.path}{the rds path}
}
\description{
instructorAssignmentModuleServer
}
