% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instructorAssignment.R
\name{instructorAssignmentModuleUI}
\alias{instructorAssignmentModuleUI}
\title{instructorAssignmentModuleUI}
\usage{
instructorAssignmentModuleUI(id)
}
\arguments{
\item{id}{The id}
}
\description{
instructorAssignmentModuleUI
}
