% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instructorLoad.R
\name{instructorLoadModuleServer}
\alias{instructorLoadModuleServer}
\title{instructorLoadModuleServer}
\usage{
instructorLoadModuleServer(
  id,
  input,
  output,
  session,
  inSemester,
  theMasterCourses,
  theCombinedData,
  theLeaveData,
  theAvailableFacultyData,
  inSemesterCodes,
  allowUpdate = FALSE,
  rds.path = NULL
)
}
\arguments{
\item{id}{The id}

\item{input}{The input}

\item{output}{The output}

\item{session}{The session}

\item{inSemester}{The input semester}

\item{theMasterCourses}{the master course list}

\item{theCombinedData}{the combined data}

\item{theLeaveData}{the leave data}

\item{theAvailableFacultyData}{the available faculty data}

\item{inSemesterCodes}{insemester codes}

\item{allowUpdate}{allow update}

\item{rds.path}{rds path}
}
\description{
instructorLoadModuleServer
}
