% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{outputModuleServer}
\alias{outputModuleServer}
\title{outputModuleServer}
\usage{
outputModuleServer(id, input, output, session, ImProxy)
}
\arguments{
\item{id}{the id}

\item{input}{the input}

\item{output}{the output}

\item{session}{the session}

\item{ImProxy}{the improxy}
}
\description{
outputModuleServer
}
