% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportFaculty.R
\name{reportFacultyModuleServer}
\alias{reportFacultyModuleServer}
\title{reportFacultyModuleServer}
\usage{
reportFacultyModuleServer(
  id,
  input,
  output,
  session,
  inSemester,
  theMasterCourses,
  theCombinedData,
  theLeaveData,
  inSemesterCodes,
  inSelectedFaculty,
  allowUpdate = FALSE,
  rds.path = NULL
)
}
\arguments{
\item{id}{the id}

\item{input}{the input}

\item{output}{the output}

\item{session}{the session}

\item{inSemester}{in semester}

\item{theMasterCourses}{the master course list}

\item{theCombinedData}{the combined data}

\item{theLeaveData}{the leave data}

\item{inSemesterCodes}{in semester codes}

\item{inSelectedFaculty}{in selected faculty}

\item{allowUpdate}{allowupdate}

\item{rds.path}{rds path}
}
\description{
reportFacultyModuleServer
}
