% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportGraphicalSchedule.R
\name{reportGraphicalScheduleModuleServer}
\alias{reportGraphicalScheduleModuleServer}
\title{reportGraphicalScheduleModuleServer}
\usage{
reportGraphicalScheduleModuleServer(
  id,
  input,
  output,
  session,
  inSemester,
  theMasterCourses,
  theCombinedData,
  theLeaveData,
  theCourseInventory,
  inSemesterCodes
)
}
\arguments{
\item{id}{the id}

\item{input}{the input}

\item{output}{the output}

\item{session}{the session}

\item{inSemester}{inSemester}

\item{theMasterCourses}{the master course list}

\item{theCombinedData}{the combined data}

\item{theLeaveData}{the leave data}

\item{theCourseInventory}{the course inventory}

\item{inSemesterCodes}{the semester codes}
}
\description{
reportGraphicalScheduleModuleServer
}
