% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportGraphicalSchedule.R
\name{reportGraphicalScheduleModuleUI}
\alias{reportGraphicalScheduleModuleUI}
\title{reportGraphicalScheduleModuleUI}
\usage{
reportGraphicalScheduleModuleUI(id)
}
\arguments{
\item{id}{the id}
}
\description{
reportGraphicalScheduleModuleUI
}
