% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportLong.R
\name{reportLongModuleServer}
\alias{reportLongModuleServer}
\title{reportLongModuleServer}
\usage{
reportLongModuleServer(
  id,
  input,
  output,
  session,
  inSemester,
  theMasterCourses,
  theCombinedData,
  theLeaveData,
  inSemesterCodes
)
}
\arguments{
\item{id}{the id}

\item{input}{the input}

\item{output}{the output}

\item{session}{the session}

\item{inSemester}{semester passed in}

\item{theMasterCourses}{the master list of courses}

\item{theCombinedData}{the combined Data}

\item{theLeaveData}{the leave data}

\item{inSemesterCodes}{all semeser codes}
}
\description{
reportLongModuleServer
}
