% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{reportModuleServer}
\alias{reportModuleServer}
\title{reportModuleServer}
\usage{
reportModuleServer(
  id,
  input,
  output,
  session,
  inSemester,
  theMasterCourses,
  theCombinedData,
  theLeaveData,
  inSemesterCodes,
  synchronize = NULL,
  chosenSemester = NULL
)
}
\arguments{
\item{id}{The id}

\item{input}{The input}

\item{output}{The output}

\item{session}{The session}

\item{inSemester}{semester passed into function}

\item{theMasterCourses}{the master list of courses}

\item{theCombinedData}{the combined data}

\item{theLeaveData}{leave data}

\item{inSemesterCodes}{all semester codes}

\item{synchronize}{whether or not the data are synchronized}

\item{chosenSemester}{which semester is chosen}
}
\description{
reportModuleServer
}
