% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportMultiSection.R
\name{reportMultiSectionModuleServer}
\alias{reportMultiSectionModuleServer}
\title{reportMultiSectionModuleServer}
\usage{
reportMultiSectionModuleServer(
  id,
  input,
  output,
  session,
  inSemester,
  theMasterCourses,
  theCombinedData,
  theLeaveData,
  inSemesterCodes
)
}
\arguments{
\item{id}{the id}

\item{input}{the input}

\item{output}{the output}

\item{session}{the session}

\item{inSemester}{input semester}

\item{theMasterCourses}{the master course list}

\item{theCombinedData}{the combined data}

\item{theLeaveData}{the leave data}

\item{inSemesterCodes}{the semester codes}
}
\description{
reportMultiSectionModuleServer
}
