% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportMultiSection.R
\name{reportMultiSectionModuleUI}
\alias{reportMultiSectionModuleUI}
\title{reportMultiSectionModuleUI}
\usage{
reportMultiSectionModuleUI(id)
}
\arguments{
\item{id}{the id}
}
\description{
reportMultiSectionModuleUI
}
