% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportNotes.R
\name{reportNotesModuleServer}
\alias{reportNotesModuleServer}
\title{reportNotesModuleServer}
\usage{
reportNotesModuleServer(
  id,
  input,
  output,
  session,
  inSemester,
  theNotes,
  theCombinedData,
  allowUpdate = FALSE,
  rds.path = NULL
)
}
\arguments{
\item{id}{the id}

\item{input}{the input}

\item{output}{the output}

\item{session}{the session}

\item{inSemester}{input semester}

\item{theNotes}{the notes}

\item{theCombinedData}{the combined data}

\item{allowUpdate}{allow updates logical}

\item{rds.path}{the rds path}
}
\description{
reportNotesModuleServer
}
