% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportUnassignedCourses.R
\name{reportUnassignedCoursesModuleServer}
\alias{reportUnassignedCoursesModuleServer}
\title{reportUnassignedCoursesModuleServer}
\usage{
reportUnassignedCoursesModuleServer(
  id,
  input,
  output,
  session,
  inSemester,
  theMasterCourses,
  theCombinedData,
  theLeaveData,
  inSemesterCodes,
  synchronize,
  chosenSemester = NULL,
  suspendWhenHidden
)
}
\arguments{
\item{id}{the id}

\item{input}{the input}

\item{output}{the output}

\item{session}{the session}

\item{inSemester}{input semester}

\item{theMasterCourses}{the master course list}

\item{theCombinedData}{the combined data}

\item{theLeaveData}{the leave data}

\item{inSemesterCodes}{the semester codes}

\item{synchronize}{synchronize boxes}

\item{chosenSemester}{the chosen semester}

\item{suspendWhenHidden}{suspend when hidden}
}
\description{
reportUnassignedCoursesModuleServer
}
