% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportUnassignedCourses.R
\name{reportUnassignedCoursesModuleUI}
\alias{reportUnassignedCoursesModuleUI}
\title{reportUnassignedCoursesModuleUI}
\usage{
reportUnassignedCoursesModuleUI(id)
}
\arguments{
\item{id}{the id}
}
\description{
reportUnassignedCoursesModuleUI
}
