% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullSchedulingBox.R
\name{fullSchedulingBoxModuleServer}
\alias{fullSchedulingBoxModuleServer}
\title{fullSchedulingBoxModuleServer}
\usage{
fullSchedulingBoxModuleServer(id, input, output, session, schedulingDataBundle)
}
\arguments{
\item{id}{The element id}

\item{input}{The input element}

\item{output}{The output element}

\item{session}{The session}

\item{schedulingDataBundle}{The entire dataset}
}
\description{
fullSchedulingBoxModuleServer
}
